Fix bug with mole context
I forgot that the concept of metadata as durability only makes sense on items, which actually have a durability. Under certain circumstances the old code failed because of that. This does not happen any more. This is actually only a temporary fix since I want to rewrite the whole MoleContext anyways.